# src/main/resources/api.yml
openapi: 3.0.3
info:
  title: Restaurant BFF OpenAPI document
  version: 0.0.1
paths:
...
  /employee/{employeeId}:
    get:
      summary: Get Employee By id.
      description: Retrieves an employee by passing an id.
      operationId: getEmployee
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Employee got."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeDTO"
    put:
      summary: Edit Employee By id.
      description: Edit employee by passing an id.
      operationId: editEmployee
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeDTO"
      responses:
        "200":
          description: "Restaurant edited."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeDTO"
    delete:
      summary: Delete Employee By id.
      description: Retrieves the deleted employee.
      operationId: deleteEmployee
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Employee deleted."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeDTO"
  /employee:
    post:
      summary: Create Employee.
      description: Create Employee.
      operationId: createEmployee
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeDTO"
      responses:
        "200":
          description: "Employee created."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeDTO"
    get:
      summary: Get All Employees.
      description: Get All Employees.
      operationId: getAllEmployees
      responses:
        "200":
          description: "Employees got."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmployeeDTO"
  /employee/add/restaurant/{employeeId}/{restaurantId}:
    put:
      summary: Add Restaurant to Employee.
      description: Add Restaurant to Employee.
      operationId: addRestaurant
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Employee with restaurant added."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeDTO"
  /employee/remove/restaurant/{employeeId}/{restaurantId}:
    put:
      summary: Remove Restaurant from Employee.
      description: Remove Restaurant from Employee.
      operationId: removeRestaurant
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Employee with restaurant removed."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeDTO"
  # LOCATION API
...
  /address/{addressId}:
    get:
      summary: Get Address By id.
      description: Retrieves a address by passing an id.
      operationId: getAddress
      parameters:
        - name: addressId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Address got."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddressDTO"
    put:
      summary: Edit Address By id.
      description: Edit address by passing an id.
      operationId: editAddress
      parameters:
        - name: addressId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressDTO"
      responses:
        "200":
          description: "Address edited."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddressDTO"
    delete:
      summary: Delete Address By id.
      description: Retrieves the deleted address.
      operationId: deleteAddress
      parameters:
        - name: addressId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Address deleted."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddressDTO"
  /address:
    post:
      summary: Create Address.
      description: Create Address.
      operationId: createAddress
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressDTO"
      responses:
        "200":
          description: "Address created."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddressDTO"
    get:
      summary: Get All Addresses.
      description: Get All Addresses.
      operationId: getAllAddresses
      responses:
        "200":
          description: "Addresses got."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AddressDTO"
servers:
  - url: http://localhost:8080/api
components:
  schemas:
    AddressDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        postalCode:
          type: string
        city:
          type: string
        country:
          type: string
...
    EmployeeDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        role:
          type: string
        restaurantList:
          type: setOfRestaurantDTO
        schedule:
          $ref: "#/components/schemas/ScheduleRoutineDTO"
...
    ScheduleRoutineDTO:
      type: object
      properties:
        daysScheduleMap:
          additionalProperties:
            type: daysScheduleMap
...